Complex types - mixed list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,12 +5,12 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> int:
+def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
     """
         :param: mxd_lst: List of floats
         :return: Sum of all numbers in list
     """
-    total_sum: int = 0
+    total_sum: float = 0.0
     for n in mxd_lst:
         total_sum = total_sum + n
